requirements.txt
cffi
characteristic
cryptography
cssselect
django
docopt
enum34
feedparser
lxml
pyquery
requests
mysql
ics
time
